Refactor system status API to use service layer
